Run entrypoint in DEV only on first creation

+++ services/backend/services/v4/entrypoints/db_migration.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+[ -e ".finished" ] && return 0
+
 echo "MONGODB_URI=$MONGODB_URI" >.env
 npm run migrate:db:up
 rm .env
